sir-ai/model-capacity-and-skill/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
PyYAML==6.0.2
jsonschema==4.23.0
pytest==8.3.4
httpx==0.28.1
